requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-filter==2.4.0
gunicorn==20.1.0
requests==2.26.0
